Added like counter function
